Expand on some points in the proposal, and reformat the markdown.
This is progress toward #1.

We don't have our consulted staff member written down on the proposal, so I've fixed that.

Our description is pretty vague at the moment. We just mention what CPU we're targeting,
and mention that we've been inspired by the sms32 simulator. I'm expanding on the points
we made in there.

Our markdown is a bit of a mess, the line lengths are really inconsistent, and I hate
inconsistency. I'm fixing that.

There's not much written under software platform, so I'm adding that we're using a Linux box
for the server, and expanding on the point about web browsers.