Reword some points in the features section, and re-do division of work
There are some weird inconsistencies in this section, we are referring to users
as both users and students, so now I'm changing all the 'students' to 'users'.

The architecture diagram will only cover the implementation, so I'm including
information about the non-implementation work.

The line lengths are inconsistent with the rest of the file. You already know
my position on this.